feat(tag): convert Tag to TSX rather than Card to Astro
Previously we converted Card to Astro because we wanted to use the Tag astro component therein, and
you are not allowed to use Astro components in React components. Unfortunately, the Card component
is used in the Search React component, which because of its advanced interactivity uses useEffect
etc which is much harder to do in Astro, thus must remain a React component. Since Card must be
imported in Search, Card needs to be a  React component, and so Tag needs to be changed to be a
React component instead.